Allow users to add videos to tournaments
Any logged in user can add a video to a tournament. They must enter a Youtube video ID (Youtube is all that is supported for now) and the record is created, and title, thumbnail and channel name are populated from the Youtube API.

The video creator, the tournament organiser or any admin can delete the video record.

When a user clicks the thumbnail or title on the tournament page, an embedded player is loaded. The player can also be closed.