Only display four videos by default
The page will only display four videos on load, hiding the others. The user can click to show/hide the excess videos. If there are fewer than five videos, the toggles are not displayed.

This should prevent the page from being overwhelmed when a lot of videos are added to a tournament.